+++ doc/source/changelog.rst
@@ -14,6 +14,7 @@ master: 1.7.1
 -------------
 
 * Expose protected members in Python derived classes
+* Fix by-ptr return of objects in Python derived classes
 * Fix for passing numpy boolean array through ``bool*``